Refactor the code and put functions in utilities.js file
