+++ src/module/dom.js
@@ -1,4 +1,5 @@
 import Storage from './localStorage.js';
+import { createLiTodo } from './utilities.js';
 
 const toDoContainer = document.querySelector('ul');
 
@@ -6,6 +7,7 @@ const updateDom = () => {
   // reset container
   toDoContainer.innerHTML = '';
   let index = 0;
+  // Check if the storage exists
   if (Storage.isEmpty()) {
     Storage.set([]);
   }
@@ -15,23 +17,12 @@ const updateDom = () => {
     // Reset the indexes starting from 1
     index += 1;
     elem.index = index;
+
     // Create li, check box and description h5 for every Todo Object
-    const li = document.createElement('li');
-    const checkbox = document.createElement('input');
-    const description = document.createElement('h6');
-    const dragableI = document.createElement('i');
-    checkbox.type = 'checkbox';
-    description.innerText = elem.description;
-    // Add classes
-    toDoContainer.classList = 'd-flex flex-column p-0';
-    li.classList += 'list-unstyled d-flex bottom-border w-100 position-relative';
-    checkbox.classList += 'my-3';
-    description.classList += 'mx-3 margin-desc';
-    dragableI.classList = 'drag-bg-img';
-    // Append check box and descripton on li
-    li.append(checkbox, description, dragableI);
-    // Append li on the ul
-    toDoContainer.appendChild(li);
+    const newLi = createLiTodo(elem.description);
+    // Append li on the ul and add classes
+    toDoContainer.classList += 'p-0 m-0';
+    toDoContainer.append(newLi);
   });
 };
